requirements.txt
fastapi
uvicorn
pydantic
numpy
pandas
yfinance
ta
torch
scikit-learn
python-binance
requests
